- вынес slider в отдельный компонент и исправил стили под картинки
